extract linkchecker into its own project
