--- linkChecker/.gitignore
@@ -1,7 +1,0 @@
-/.settings/
-/.project
-.idea/
-*.iml
-/target/
-.classpath
-/bin/
