--- linkChecker/src/main/resources/file.properties
@@ -1,23 +1,0 @@
-DATABASE_NAME=links
-
-#if this is empty, it will try localhost:27017
-DATABASE_URI=
-
-#in milliseconds
-TIMEOUT=5000
-REDIRECT_FOLLOW_LIMIT=5
-USERAGENT=CLARIN Curation LinkChecker@acdh.oeaw.ac.at
-
-#in milliseconds
-CRAWLDELAY=0
-
-#crawldelaylist is a list for specific collections
-#entries are divided by comma and label value pairs are divided by the equal sign
-#example:CRAWLDELAYLIST=The_Language_Archive=10000,BAS_Repository=2000
-#for now we only have "The_Language_Archive"
-CRAWLDELAYLIST=The_Language_Archive=10000
-
-#this parameter causes linkchecker to copy all broken urls from linksChecked to
-#linksToBeChecked at the start of the program (only once)
-#to put it into effect, you have to start the program again
-ONLY_BROKEN=true
