--- linkChecker/src/test/resources/log4j.properties
@@ -1,19 +1,0 @@
-
-log4j.rootLogger=INFO
-
-log4j.appender.Stdout=org.apache.log4j.ConsoleAppender
-log4j.appender.Stdout.layout=org.apache.log4j.PatternLayout
-log4j.appender.Stdout.layout.conversionPattern=%d %p [%c{1}#%M:%L] - %m%n
-
-log4j.appender.File.File=${catalina.base}/logs/curate.log
-log4j.appender.File=org.apache.log4j.DailyRollingFileAppender
-log4j.appender.File.layout=org.apache.log4j.PatternLayout
-log4j.appender.File.layout.ConversionPattern=%d %p [%c#%M:%L] - %m%n
-
-log4j.appender.R=org.apache.log4j.RollingFileAppender
-log4j.appender.R.File=dump.out
-log4j.appender.R.layout=org.apache.log4j.PatternLayout
-log4j.appender.R.layout.ConversionPattern=%p %t %c - %m%n
-
-log4j.logger.org.apache=WARN
-log4j.logger.httpclient=WARN
